refactor: move dataset formatting helpers to lib/
